Add content, links to index.html
This landing page will contain information about me as well as links to
my github and other sites. It is meant as a professional landing page
where users can go to my portfolio (github), selected projects (like
quake map), and visit other links, such as my photography site.

This is a first draft and will continue to be updated and polished.

* Add about info
* Add links to github, quake map, linkedin, and my art photography
  website.